+++ src/events/ready.js
@@ -2,6 +2,7 @@ const { Events } = require('discord.js');
 const { startReminderScheduler } = require('../utils/reminderScheduler');
 const { startGiveawayScheduler } = require('../utils/giveawayScheduler');
 const { startBirthdayScheduler } = require('../utils/birthdayScheduler');
+const { startAlertScheduler } = require('../utils/alertScheduler');
 const { startHeartbeatMonitor } = require('../utils/heartbeat');
 
 module.exports = {
@@ -12,6 +13,7 @@ module.exports = {
     startReminderScheduler(client);
     startGiveawayScheduler(client);
     startBirthdayScheduler(client);
+    startAlertScheduler(client);
     startHeartbeatMonitor(client);
   },
 };